+++ lucid/models/vision/inception_next/__init__.py
@@ -6,8 +6,17 @@ from lucid.models.vision.inception_next._model import (
     InceptionNeXtForImageClassification,
 )
 from lucid.models.vision.inception_next._pretrained import (
+    inception_next_base,
+    inception_next_base_cls,
+    inception_next_small,
+    inception_next_small_cls,
     inception_next_tiny,
     inception_next_tiny_cls,
+)
+from lucid.models.vision.inception_next._weights import (
+    InceptionNeXtBaseWeights,
+    InceptionNeXtSmallWeights,
+    InceptionNeXtTinyWeights,
 )
 
 __all__ = [
@@ -16,4 +25,11 @@ __all__ = [
     "InceptionNeXtForImageClassification",
     "inception_next_tiny",
     "inception_next_tiny_cls",
+    "inception_next_small",
+    "inception_next_small_cls",
+    "inception_next_base",
+    "inception_next_base_cls",
+    "InceptionNeXtTinyWeights",
+    "InceptionNeXtSmallWeights",
+    "InceptionNeXtBaseWeights",
 ]